Refactor ddPCR to dPCR
Note that this commit introduces breaking changes for all model component functions related to d(d)PCR. The refactoring reflects that there are further digital PCR (dPCR) methods besides digital droplet PCR (ddPCR) that are not based on droplets but other kinds of partitions. Since the measurement noise and LOD model applies to all forms of dPCR, the naming of functions and documentation has been adapted to be more generic.