Implement workaround to avoid Rt bias at very small infection numbers
This is achieved by adding a soft upper bound on the coefficient of variation of new infections, to avoid a bias in the mean of the truncated normal distribution, which in turn biases Rt downwards otherwise. Only relevant for very small infection numbers, i.e. less than 3 infections per day